Trace Pub/Sub publishing
The Pub/Sub client publishes messages by placing them onto a client
buffer and having a bundler (working asynchronously) pull them from the
buffer and submit them in batch to GCP.

This means there's a disconnect from our app publishing messages and the
process that publishes them. As we're seeing timeouts around publishing
we want to connect the app publish to the bundler publish.

We achieve this by:

1. Adding tracing around flow control so waiting on a topic semaphore is
   exposed in our traces
2. Creates spans that get linked between the application publish and the
   bundler publish that allow inspecting the gRPC operation that
   actually submits things to Pub/Sub

^[3]: https://github.com/open-telemetry/opentelemetry-go/pull/5032

This isn't as nice as it could be because OTEL does not yet support
adding links to spans after they've been created (see comment for more
detail). We will be able to clean this up once that feature lands (see
[3]).